Fix "ReferenceError: Can't find variable: segments" masking JSON parsing errors.
This appears to have been a copy-paste error in 6b97aa2 and caused rendering of schemas that triggered such parsing errors to be aborted before anything was rendered.

Also output the content that failed to parse for more info.

This brings test.html output back to what is currently seen at http://lbovet.github.io/docson/tests/test.html, although with a wrongful such JSON parsing error on the console, to be fixed in the next commit.