Fix attempt to parse the HTML page as JSON for the root schema.
In schemas that contained internal references ({"$ref": "#..."}), Docson would attempt to retrieve the document named by that URI using $.get(). That would result in the HTML code for the page it was running in, which is completely unrelated and would fail to parse as JSON, resulting in an error message on the console.

To avoid attempting to retrieve the root document again, initialize the document collection with it already.

As a side effect, this causes a circular reference to the root schema to be detected at the first cycle rather than the second, which makes integration test "recursive schema" fail, but actually seems desirable to me.